Fix: 스테이지 맵 수정

+++ Assets/seocxwoo/Scripts/Ingredient/Ingredient.cs
@@ -14,7 +14,7 @@ public class Ingredient : MonoBehaviour
         return data.height;
     }
 
-    public float GetIndex()
+    public int GetIndex()
     {
         return data.index;
     }
